+++ src/crossover/twoPoint.js
@@ -1,5 +1,5 @@
 const kPoint = require('./kPoint');
 
-const twoPoint = kPoint(2);
+const twoPoint = () => kPoint(2);
 
 module.exports = twoPoint;