+++ src/evaluatePopulation.js
@@ -1,0 +1,7 @@
+const evaluatePopulation = ({ fitnessFunction }) => population => (
+  population.map(individual => (
+    fitnessFunction(individual)
+  ))
+);
+
+module.exports = evaluatePopulation;