+++ src/index.js
@@ -1,11 +1,12 @@
-const { runEvolution, runEvolutionAsync, getGenerationsIterator } = require('./runners');
+const { run } = require('./runners');
 
 const selection = require('./selection');
 const mutation = require('./mutation');
 const crossover = require('./crossover');
 
 const { generateInitialPopulation } = require('./generateInitialPopulation');
-const { reproduce, reproduceAsync } = require('./reproduce');
+const evaluatePopulation = require('./evaluatePopulation');
+const { reproduce } = require('./reproduce');
 const { stopCondition } = require('./stopConditions');
 const randomSequenceOf = require('./randomSequenceOf');
 const randomPermutationOf = require('./randomPermutationOf');
@@ -13,17 +14,15 @@ const elitism = require('./elitism');
 const logIterationData = require('./logIterationData');
 
 module.exports = {
-  runEvolution,
-  runEvolutionAsync,
-  getGenerationsIterator,
+  run,
 
   selection,
   mutation,
   crossover,
 
   generateInitialPopulation,
+  evaluatePopulation,
   reproduce,
-  reproduceAsync,
   stopCondition,
   logIterationData,
   randomSequenceOf,