+++ test/integration/findSequence.test.js
@@ -2,7 +2,7 @@ const rand = require('random-seed');
 const Genemo = require('../../src');
 
 describe('Find Sequence', () => {
-  test('Last generation matches snapshot', () => {
+  test('Last iteration population matches snapshot', async (done) => {
     const { random } = rand.create('seed');
 
     const generateIndividual = Genemo.randomSequenceOf([true, false], 50);
@@ -19,18 +19,20 @@ describe('Find Sequence', () => {
         generateIndividual,
         size: 50,
       }),
-      selection: Genemo.selection.roulette(),
+      selection: Genemo.selection.roulette({ minimizeFitness: false }),
       reproduce: Genemo.reproduce({
-        crossover: Genemo.crossover.singlePoint,
-        mutate: Genemo.mutation.transformRandomGene(Genemo.mutation.flipBit),
+        crossover: Genemo.crossover.singlePoint(),
+        mutate: Genemo.mutation.transformRandomGene(Genemo.mutation.flipBit()),
         mutationProbability: 0.03,
       }),
-      fitness: fitnessFunction,
-      stopCondition: Genemo.stopCondition({ minFitness: 50, maxGenerations: 100 }),
+      evaluatePopulation: Genemo.evaluatePopulation({ fitnessFunction }),
+      stopCondition: Genemo.stopCondition({ minFitness: 50, maxIterations: 100 }),
       random,
+      collectLogs: false,
     };
 
-    const lastGeneration = Genemo.runEvolution(evolutionOptions);
-    expect(lastGeneration).toMatchSnapshot('findSequence1');
+    const result = await Genemo.run(evolutionOptions);
+    expect(result).toMatchSnapshot('findSequence1');
+    done();
   });
 });