+++ test/stopConditions.test.js
@@ -7,16 +7,16 @@ describe('stopCondition', () => {
       { individual: 2, fitness: 5 },
       { individual: 3, fitness: 0 },
     ];
-    const generation = 50;
+    const iteration = 50;
 
-    const stopCondition1 = Genemo.stopCondition({ minFitness: 10, maxGenerations: 100 });
-    const stopCondition2 = Genemo.stopCondition({ minFitness: 4, maxGenerations: 100 });
+    const stopCondition1 = Genemo.stopCondition({ minFitness: 10, maxIterations: 100 });
+    const stopCondition2 = Genemo.stopCondition({ minFitness: 4, maxIterations: 100 });
     const stopCondition3 = Genemo.stopCondition({ maxFitness: 10 });
-    const stopCondition4 = Genemo.stopCondition({ maxGenerations: 10 });
+    const stopCondition4 = Genemo.stopCondition({ maxIterations: 10 });
 
-    expect(stopCondition1({ evaluatedPopulation, generation })).toBe(false);
-    expect(stopCondition2({ evaluatedPopulation, generation })).toBe(true);
-    expect(stopCondition3({ evaluatedPopulation, generation })).toBe(true);
-    expect(stopCondition4({ evaluatedPopulation, generation })).toBe(true);
+    expect(stopCondition1({ evaluatedPopulation, iteration })).toBe(false);
+    expect(stopCondition2({ evaluatedPopulation, iteration })).toBe(true);
+    expect(stopCondition3({ evaluatedPopulation, iteration })).toBe(true);
+    expect(stopCondition4({ evaluatedPopulation, iteration })).toBe(true);
   });
 });